change entity

+++ src/main/java/api/lemonico/entity/WarehouseEntity.java
@@ -48,7 +48,7 @@ public class WarehouseEntity extends LcEntity
     /** 更新日時 */
     @Column(name = "modified_at")
     LocalDateTime modifiedAt;
-    /** 削除フラグ（退会から一定時間経過後に削除状態になる） */
+    /** 削除フラグ（0: 未削除 1: 削除済） */
     @Column(name = "is_deleted")
-    Integer isDeleted;
+    Byte isDeleted;
 }
